requirements.txt
dacite
kubernetes>=28,<29
flask==2.3.2
jsonify
waitress
PyHamcrest
requests_mock
dataclasses-json
python-dotenv
pytest
git+https://github.com/ucsd-ets/awsed_python_client.git@2024.2.1-RC1
